put generic button in buttons file, fix url for merge
