+++ src/components/app-frame.tsx
@@ -12,7 +12,6 @@ import { onGetSections } from 'actions';
 import { Modals } from 'contexts/page-modals';
 import { AppServices } from 'services';
 import Loader from './loader';
-import { checker } from 'reducers/modals';
 
 const PAGE_TITLE = process.env.REACT_APP_PAGE_TITLE;
 
@@ -53,7 +52,6 @@ export interface AppFrameProps{
 }
 
 const AppFrame: React.FunctionComponent<AppFrameProps> = ({ children }) => {
-    checker();
     const userData: User = useContext(UserContext);
     const { sectionsService: service }: AppServices = useAppService();
     const dispatch = useAppDispatch();
