+++ src/components/controls-bar/buttons.tsx
@@ -1,27 +1,13 @@
 import React from 'react';
-import { Button, Theme } from '@material-ui/core';
-import { makeStyles, createStyles } from '@material-ui/styles';
+import { Button } from '@material-ui/core';
 import MergeIcon from '@material-ui/icons/MergeType';
 import AddIcon from '@material-ui/icons/Add';
 import CancelIcon from '@material-ui/icons/Cancel';
-import amber from '@material-ui/core/colors/amber';
 import { useModalsDispatch, useModalsState } from 'contexts/page-modals';
 import { onToggleMergeModal, onToggleAddModal } from 'actions';
 import { ClickHandler } from 'global-types';
+import { useButtonStyles } from 'components/buttons';
 
-export const useButtonStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        button: {
-            margin: theme.spacing(1),
-            color: theme.palette.primary.main
-        },
-        btnConfirm: {
-            backgroundColor: amber[500]
-        },
-        icon: {
-            marginLeft: theme.spacing(1)
-        }
-    }));
 
 export interface MergeButtonProps {
     onClick: ClickHandler;
@@ -83,19 +69,3 @@ export const AddSectionButton = () => {
     );
 };
 
-export interface ConfirmBtnProps {
-    onClick: ((event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void) | undefined;
-    children: React.ReactNode;
-}
-
-export const ConfirmButton: React.FC<ConfirmBtnProps> = ({ onClick, children }) => {
-    const styles = useButtonStyles();
-
-    return (
-        <Button onClick={onClick}
-            color="secondary"
-            className={styles.button}
-        >
-            {children}
-        </Button>);
-};
